Mapper updates and additions

+++ src/ppu.c
@@ -90,8 +90,8 @@ static void makeppulut(void) {
 	}
 }
 
-static int ppudead = 1;
-static int kook = 0;
+static uint8 ppudead = 1;
+static uint8 kook = 0;
 int fceuindbg = 0;
 
 int MMC5Hack = 0, PEC586Hack = 0;
